# SocialMedia.Core/DTOs/PostDto.cs
using System;
// using System.ComponentModel.DataAnnotations;

namespace SocialMedia.Core.DTOs
{
    public class PostDto
    {
        /// <summary>
        /// Autogenerated id for post entity
        /// </summary>
        public int Id { get; set; }
        public int UserId { get; set; }
        public DateTime? Date { get; set; }

        // [Required]
        public string Description { get; set; }
        public string Image { get; set; }
    }
}